// src/app/models/User.ts
export class User {
    id: number;
    status: number;
    password: string;
    email: string;
    sex: number;
    name: string;
    firstname: string;
    wallet: number;
    registration_date: string;
    phone: number;
    address: string;
    postal_code: string;
    town: string;
    image_id: number;
  }